Add sanity check CI pipeline for versioning
Ensure that on every commit, the line
`__version__ = ...  # semantic-version-placeholder`
exists in the `__init__.py` file in `lightshow`. This line
is replaced in the build script with the proper version as per
dunamai and the git tag/commit. sed is used to do the replacement
so this line must match exactly.